--- a/commit/src/pcs.rs
+++ b/commit/src/pcs.rs
@@ -6,7 +6,7 @@ use alloc::vec::Vec;
 use p3_challenger::FieldChallenger;
 use p3_field::{ExtensionField, Field};
 use p3_matrix::dense::RowMajorMatrixView;
-use p3_matrix::MatrixRows;
+use p3_matrix::{Dimensions, MatrixRows};
 
 /// A (not necessarily hiding) polynomial commitment scheme, for committing to (batches of)
 /// polynomials defined over the field `F`.
@@ -54,6 +54,7 @@ where
     fn verify_multi_batches(
         &self,
         commits_and_points: &[(Self::Commitment, &[EF])],
+        dims: &[Vec<Dimensions>],
         values: OpenedValues<EF>,
         proof: &Self::Proof,
         challenger: &mut Challenger,