--- a/dft/src/lib.rs
+++ b/dft/src/lib.rs
@@ -19,4 +19,3 @@ pub use radix_2_bowers::*;
 pub use radix_2_dit::*;
 pub use radix_2_dit_parallel::*;
 pub use traits::*;
-pub use util::*;