--- a/dft/src/radix_2_bowers.rs
+++ b/dft/src/radix_2_bowers.rs
@@ -4,14 +4,13 @@ use p3_field::{Field, Powers, TwoAdicField};
 use p3_matrix::dense::{RowMajorMatrix, RowMajorMatrixViewMut};
 use p3_matrix::Matrix;
 use p3_maybe_rayon::{IndexedParallelIterator, MaybeParChunksMut, ParallelIterator};
-use p3_util::log2_strict_usize;
+use p3_util::{log2_strict_usize, reverse_bits};
 
 use crate::butterflies::{
     dif_butterfly_on_rows, dit_butterfly_on_rows, twiddle_free_butterfly_on_rows,
 };
 use crate::util::{
-    bit_reversed_zero_pad, divide_by_height, reverse_bits, reverse_matrix_index_bits,
-    reverse_slice_index_bits,
+    bit_reversed_zero_pad, divide_by_height, reverse_matrix_index_bits, reverse_slice_index_bits,
 };
 use crate::TwoAdicSubgroupDft;