--- a/dft/src/radix_2_dit.rs
+++ b/dft/src/radix_2_dit.rs
@@ -15,17 +15,25 @@ use crate::TwoAdicSubgroupDft;
 pub struct Radix2Dit;
 
 impl<F: TwoAdicField> TwoAdicSubgroupDft<F> for Radix2Dit {
-    fn dft_batch(&self, mut mat: RowMajorMatrix<F>) -> RowMajorMatrix<F> {
+    fn dft_batch_bitrev<const IN_BITREV: bool, const OUT_BITREV: bool>(
+        &self,
+        mut mat: RowMajorMatrix<F>,
+    ) -> RowMajorMatrix<F> {
         let h = mat.height();
         let log_h = log2_strict_usize(h);
 
         let root = F::two_adic_generator(log_h);
         let twiddles: Vec<F> = root.powers().take(h / 2).collect();
 
+        if !IN_BITREV {
+            reverse_matrix_index_bits(&mut mat);
+        }
         // DIT butterfly
-        reverse_matrix_index_bits(&mut mat);
         for layer in 0..log_h {
             dit_layer(&mut mat.as_view_mut(), layer, &twiddles);
+        }
+        if OUT_BITREV {
+            reverse_matrix_index_bits(&mut mat);
         }
         mat
     }