--- a/dft/src/radix_2_dit.rs
+++ b/dft/src/radix_2_dit.rs
@@ -2,12 +2,12 @@ use alloc::vec::Vec;
 
 use p3_field::{Field, TwoAdicField};
 use p3_matrix::dense::{RowMajorMatrix, RowMajorMatrixViewMut};
+use p3_matrix::util::reverse_matrix_index_bits;
 use p3_matrix::Matrix;
 use p3_maybe_rayon::{IndexedParallelIterator, MaybeParChunksMut, ParallelIterator};
 use p3_util::log2_strict_usize;
 
 use crate::butterflies::{dit_butterfly_on_rows, twiddle_free_butterfly_on_rows};
-use crate::util::reverse_matrix_index_bits;
 use crate::TwoAdicSubgroupDft;
 
 /// The DIT FFT algorithm.
@@ -15,25 +15,19 @@ use crate::TwoAdicSubgroupDft;
 pub struct Radix2Dit;
 
 impl<F: TwoAdicField> TwoAdicSubgroupDft<F> for Radix2Dit {
-    fn dft_batch_bitrev<const IN_BITREV: bool, const OUT_BITREV: bool>(
-        &self,
-        mut mat: RowMajorMatrix<F>,
-    ) -> RowMajorMatrix<F> {
+    type Evaluations = RowMajorMatrix<F>;
+
+    fn dft_batch(&self, mut mat: RowMajorMatrix<F>) -> RowMajorMatrix<F> {
         let h = mat.height();
         let log_h = log2_strict_usize(h);
 
         let root = F::two_adic_generator(log_h);
         let twiddles: Vec<F> = root.powers().take(h / 2).collect();
 
-        if !IN_BITREV {
-            reverse_matrix_index_bits(&mut mat);
-        }
+        reverse_matrix_index_bits(&mut mat);
         // DIT butterfly
         for layer in 0..log_h {
             dit_layer(&mut mat.as_view_mut(), layer, &twiddles);
-        }
-        if OUT_BITREV {
-            reverse_matrix_index_bits(&mut mat);
         }
         mat
     }