--- a/dft/src/radix_2_dit_parallel.rs
+++ b/dft/src/radix_2_dit_parallel.rs
@@ -4,11 +4,11 @@ use p3_field::{Field, Powers, TwoAdicField};
 use p3_matrix::dense::{RowMajorMatrix, RowMajorMatrixViewMut};
 use p3_matrix::Matrix;
 use p3_maybe_rayon::{IndexedParallelIterator, ParallelIterator};
-use p3_util::log2_strict_usize;
+use p3_util::{log2_strict_usize, reverse_bits};
 
 use crate::butterflies::dit_butterfly;
 use crate::util::{bit_reversed_zero_pad, reverse_matrix_index_bits};
-use crate::{reverse_bits, reverse_slice_index_bits, TwoAdicSubgroupDft};
+use crate::{reverse_slice_index_bits, TwoAdicSubgroupDft};
 
 /// A parallel FFT algorithm which divides a butterfly network's layers into two halves.
 ///
@@ -22,6 +22,11 @@ pub struct Radix2DitParallel;
 
 impl<F: TwoAdicField> TwoAdicSubgroupDft<F> for Radix2DitParallel {
     fn dft_batch(&self, mut mat: RowMajorMatrix<F>) -> RowMajorMatrix<F> {
+        mat = self.dft_batch_bit_reversed(mat);
+        reverse_matrix_index_bits(&mut mat);
+        mat
+    }
+    fn dft_batch_bit_reversed(&self, mut mat: RowMajorMatrix<F>) -> RowMajorMatrix<F> {
         let h = mat.height();
         let log_h = log2_strict_usize(h);
 
@@ -38,12 +43,22 @@ impl<F: TwoAdicField> TwoAdicSubgroupDft<F> for Radix2DitParallel {
         reverse_matrix_index_bits(&mut mat);
         reverse_slice_index_bits(&mut twiddles);
         par_dit_layer_rev(&mut mat, mid, &twiddles);
-        reverse_matrix_index_bits(&mut mat);
 
         mat
     }
 
     fn coset_lde_batch(
+        &self,
+        mut mat: RowMajorMatrix<F>,
+        added_bits: usize,
+        shift: F,
+    ) -> RowMajorMatrix<F> {
+        mat = self.coset_lde_batch_bit_reversed(mat, added_bits, shift);
+        reverse_matrix_index_bits(&mut mat);
+        mat
+    }
+
+    fn coset_lde_batch_bit_reversed(
         &self,
         mut mat: RowMajorMatrix<F>,
         added_bits: usize,
@@ -99,7 +114,6 @@ impl<F: TwoAdicField> TwoAdicSubgroupDft<F> for Radix2DitParallel {
         reverse_matrix_index_bits(&mut mat);
         reverse_slice_index_bits(&mut twiddles);
         par_dit_layer_rev(&mut mat, mid, &twiddles);
-        reverse_matrix_index_bits(&mut mat);
 
         mat
     }