--- a/dft/src/radix_2_dit_parallel.rs
+++ b/dft/src/radix_2_dit_parallel.rs
@@ -1,13 +1,15 @@
 use alloc::vec::Vec;
 
 use p3_field::{Field, Powers, TwoAdicField};
+use p3_matrix::bitrev::BitReversedMatrixView;
 use p3_matrix::dense::{RowMajorMatrix, RowMajorMatrixViewMut};
+use p3_matrix::util::reverse_matrix_index_bits;
 use p3_matrix::Matrix;
 use p3_maybe_rayon::{IndexedParallelIterator, ParallelIterator};
 use p3_util::{log2_strict_usize, reverse_bits, reverse_slice_index_bits};
 
 use crate::butterflies::dit_butterfly;
-use crate::util::{bit_reversed_zero_pad, reverse_matrix_index_bits};
+use crate::util::bit_reversed_zero_pad;
 use crate::TwoAdicSubgroupDft;
 
 /// A parallel FFT algorithm which divides a butterfly network's layers into two halves.
@@ -21,10 +23,9 @@ use crate::TwoAdicSubgroupDft;
 pub struct Radix2DitParallel;
 
 impl<F: TwoAdicField> TwoAdicSubgroupDft<F> for Radix2DitParallel {
-    fn dft_batch_bitrev<const IN_BITREV: bool, const OUT_BITREV: bool>(
-        &self,
-        mut mat: RowMajorMatrix<F>,
-    ) -> RowMajorMatrix<F> {
+    type Evaluations = BitReversedMatrixView<RowMajorMatrix<F>>;
+
+    fn dft_batch(&self, mut mat: RowMajorMatrix<F>) -> Self::Evaluations {
         let h = mat.height();
         let log_h = log2_strict_usize(h);
 
@@ -34,28 +35,23 @@ impl<F: TwoAdicField> TwoAdicSubgroupDft<F> for Radix2DitParallel {
         let mid = log_h / 2;
 
         // The first half looks like a normal DIT.
-        if !IN_BITREV {
-            reverse_matrix_index_bits(&mut mat);
-        }
+        reverse_matrix_index_bits(&mut mat);
         par_dit_layer(&mut mat, mid, &twiddles);
 
         // For the second half, we flip the DIT, working in bit-reversed order.
         reverse_matrix_index_bits(&mut mat);
         reverse_slice_index_bits(&mut twiddles);
         par_dit_layer_rev(&mut mat, mid, &twiddles);
-        if !OUT_BITREV {
-            reverse_matrix_index_bits(&mut mat);
-        }
 
-        mat
+        BitReversedMatrixView::new(mat)
     }
 
-    fn coset_lde_batch_bitrev<const IN_BITREV: bool, const OUT_BITREV: bool>(
+    fn coset_lde_batch(
         &self,
         mut mat: RowMajorMatrix<F>,
         added_bits: usize,
         shift: F,
-    ) -> RowMajorMatrix<F> {
+    ) -> Self::Evaluations {
         let h = mat.height();
         let log_h = log2_strict_usize(h);
         let mid = log_h / 2;
@@ -67,9 +63,7 @@ impl<F: TwoAdicField> TwoAdicSubgroupDft<F> for Radix2DitParallel {
         let mut twiddles_inv: Vec<F> = root_inv.powers().take(h / 2).collect();
 
         // The first half looks like a normal DIT.
-        if !IN_BITREV {
-            reverse_matrix_index_bits(&mut mat);
-        }
+        reverse_matrix_index_bits(&mut mat);
         par_dit_layer(&mut mat, mid, &twiddles_inv);
 
         // For the second half, we flip the DIT, working in bit-reversed order.
@@ -109,11 +103,7 @@ impl<F: TwoAdicField> TwoAdicSubgroupDft<F> for Radix2DitParallel {
         reverse_slice_index_bits(&mut twiddles);
         par_dit_layer_rev(&mut mat, mid, &twiddles);
 
-        if !OUT_BITREV {
-            reverse_matrix_index_bits(&mut mat);
-        }
-
-        mat
+        BitReversedMatrixView::new(mat)
     }
 }