--- a/dft/src/testing.rs
+++ b/dft/src/testing.rs
@@ -1,5 +1,6 @@
 use p3_field::TwoAdicField;
 use p3_matrix::dense::RowMajorMatrix;
+use p3_matrix::MatrixRows;
 use rand::distributions::{Distribution, Standard};
 use rand::thread_rng;
 
@@ -18,7 +19,7 @@ where
         let mat = RowMajorMatrix::<F>::rand(&mut rng, h, 3);
         let dft_naive = NaiveDft.dft_batch(mat.clone());
         let dft_result = dft.dft_batch(mat);
-        assert_eq!(dft_naive, dft_result);
+        assert_eq!(dft_naive, dft_result.to_row_major_matrix());
     }
 }
 
@@ -36,7 +37,7 @@ where
         let shift = F::generator();
         let coset_dft_naive = NaiveDft.coset_dft_batch(mat.clone(), shift);
         let coset_dft_result = dft.coset_dft_batch(mat, shift);
-        assert_eq!(coset_dft_naive, coset_dft_result);
+        assert_eq!(coset_dft_naive, coset_dft_result.to_row_major_matrix());
     }
 }
 
@@ -70,7 +71,7 @@ where
         let mat = RowMajorMatrix::<F>::rand(&mut rng, h, 3);
         let lde_naive = NaiveDft.lde_batch(mat.clone(), 1);
         let lde_result = dft.lde_batch(mat, 1);
-        assert_eq!(lde_naive, lde_result);
+        assert_eq!(lde_naive, lde_result.to_row_major_matrix());
     }
 }
 
@@ -88,7 +89,7 @@ where
         let shift = F::generator();
         let coset_lde_naive = NaiveDft.coset_lde_batch(mat.clone(), 1, shift);
         let coset_lde_result = dft.coset_lde_batch(mat, 1, shift);
-        assert_eq!(coset_lde_naive, coset_lde_result);
+        assert_eq!(coset_lde_naive, coset_lde_result.to_row_major_matrix());
     }
 }
 
@@ -104,7 +105,7 @@ where
         let h = 1 << log_h;
         let original = RowMajorMatrix::<F>::rand(&mut rng, h, 3);
         let dft_output = dft.dft_batch(original.clone());
-        let idft_output = dft.idft_batch(dft_output);
+        let idft_output = dft.idft_batch(dft_output.to_row_major_matrix());
         assert_eq!(original, idft_output);
     }
 }