--- a/dft/src/util.rs
+++ b/dft/src/util.rs
@@ -4,7 +4,7 @@ use p3_field::Field;
 use p3_matrix::dense::RowMajorMatrix;
 use p3_matrix::Matrix;
 use p3_maybe_rayon::{MaybeIntoParIter, ParallelIterator};
-use p3_util::log2_strict_usize;
+use p3_util::{log2_strict_usize, reverse_bits_len};
 
 pub fn reverse_slice_index_bits<F>(vals: &mut [F]) {
     let n = vals.len();
@@ -34,22 +34,6 @@ pub fn reverse_matrix_index_bits<F>(mat: &mut RowMajorMatrix<F>) {
             unsafe { swap_rows_raw(values, w, i, j) };
         }
     });
-}
-
-#[inline]
-pub const fn reverse_bits(x: usize, n: usize) -> usize {
-    reverse_bits_len(x, n.trailing_zeros() as usize)
-}
-
-#[inline]
-pub const fn reverse_bits_len(x: usize, bit_len: usize) -> usize {
-    // NB: The only reason we need overflowing_shr() here as opposed
-    // to plain '>>' is to accommodate the case n == num_bits == 0,
-    // which would become `0 >> 64`. Rust thinks that any shift of 64
-    // bits causes overflow, even when the argument is zero.
-    x.reverse_bits()
-        .overflowing_shr(usize::BITS - bit_len as u32)
-        .0
 }
 
 /// Assumes `i < j`.