--- a/fri/src/lib.rs
+++ b/fri/src/lib.rs
@@ -1,6 +1,6 @@
 //! An implementation of the FRI low-degree test (LDT).
 
-#![no_std]
+// #![no_std]
 
 extern crate alloc;
 
@@ -16,6 +16,7 @@ mod config;
 mod fold_even_odd;
 mod matrix_reducer;
 mod proof;
+mod query_index;
 mod prover;
 mod verifier;