Needs cleanup!

FRI now takes bit-reversed LDE, which is needed for multi-degree to work by `reduced_idx = idx >> n`

For now we can insert bit reverses for existing `UnivariatePcsWithLde` users (`uni-stark`), but these are repeating a lot of unnecessary work, so `uni-stark` quotient computation should use a `ZeroifierOnCoset` with the proper x indices