@rjeli 2023-12-12
Single-threaded M1 Pro:

Main:
```
INFO     generate Keccak trace [ 582ms | 100.00% ]
INFO     prove [ 5.18s | 19.80% / 100.00% ]
INFO     ┝━ commit to trace data [ 1.74s | 0.00% / 33.63% ]
INFO     │  ┝━ compute all coset LDEs [ 446ms | 8.62% ]
DEBUG    │  ┕━ build merkle tree [ 1.29s | 25.01% ] dimensions: "[2733x32768]"
INFO     ┝━ compute quotient polynomial [ 620ms | 11.98% ]
INFO     ┝━ decompose quotient polynomial [ 1.71ms | 0.03% ]
INFO     ┝━ commit to quotient poly chunks [ 24.2ms | 0.00% / 0.47% ]
INFO     │  ┝━ compute all coset LDEs [ 2.41ms | 0.05% ]
DEBUG    │  ┕━ build merkle tree [ 21.8ms | 0.42% ] dimensions: "[8x32768]"
INFO     ┕━ prove batch opening [ 1.76s | 0.05% / 34.09% ]
INFO        ┝━ compute opened values with Lagrange interpolation [ 343ms | 6.62% ]
INFO        ┕━ FRI prover [ 1.42s | 0.04% / 27.41% ]
INFO           ┝━ commit phase [ 1.42s | 0.05% / 27.36% ]
DEBUG          │  ┝━ fold in matrices [ 1.39s | 26.84% ] height: 32768
DEBUG          │  ┝━ build merkle tree [ 10.9ms | 0.21% ] dimensions: "[8x16384]"
DEBUG          │  ┝━ fold_even_odd [ 1.08ms | 0.02% ]
DEBUG          │  ┝━ build merkle tree [ 5.48ms | 0.11% ] dimensions: "[8x8192]"
DEBUG          │  ┝━ fold_even_odd [ 516µs | 0.01% ]
DEBUG          │  ┝━ build merkle tree [ 2.77ms | 0.05% ] dimensions: "[8x4096]"
DEBUG          │  ┝━ fold_even_odd [ 255µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 1.41ms | 0.03% ] dimensions: "[8x2048]"
DEBUG          │  ┝━ fold_even_odd [ 132µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 718µs | 0.01% ] dimensions: "[8x1024]"
DEBUG          │  ┝━ fold_even_odd [ 67.6µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 375µs | 0.01% ] dimensions: "[8x512]"
DEBUG          │  ┝━ fold_even_odd [ 34.5µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 206µs | 0.00% ] dimensions: "[8x256]"
DEBUG          │  ┝━ fold_even_odd [ 17.8µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 124µs | 0.00% ] dimensions: "[8x128]"
DEBUG          │  ┝━ fold_even_odd [ 9.71µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 71.7µs | 0.00% ] dimensions: "[8x64]"
DEBUG          │  ┝━ fold_even_odd [ 5.62µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 37.5µs | 0.00% ] dimensions: "[8x32]"
DEBUG          │  ┝━ fold_even_odd [ 3.17µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 26.7µs | 0.00% ] dimensions: "[8x16]"
DEBUG          │  ┝━ fold_even_odd [ 2.17µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 17.0µs | 0.00% ] dimensions: "[8x8]"
DEBUG          │  ┝━ fold_even_odd [ 1.54µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 9.29µs | 0.00% ] dimensions: "[8x4]"
DEBUG          │  ┝━ fold_even_odd [ 1.42µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 4.08µs | 0.00% ] dimensions: "[8x2]"
DEBUG          │  ┕━ fold_even_odd [ 1.33µs | 0.00% ]
INFO           ┕━ query phase [ 503µs | 0.01% ]
```

This branch:
```
INFO     generate Keccak trace [ 588ms | 100.00% ]
INFO     prove [ 4.52s | 24.77% / 100.00% ]
INFO     ┝━ commit to trace data [ 1.76s | 0.00% / 38.84% ]
INFO     │  ┝━ compute all coset LDEs [ 461ms | 10.19% ]
DEBUG    │  ┕━ build merkle tree [ 1.30s | 28.65% ] dimensions: "[2733x32768]"
INFO     ┝━ compute quotient polynomial [ 607ms | 13.41% ]
INFO     ┝━ decompose quotient polynomial [ 1.75ms | 0.04% ]
INFO     ┝━ commit to quotient poly chunks [ 24.9ms | 0.00% / 0.55% ]
INFO     │  ┝━ compute all coset LDEs [ 3.19ms | 0.07% ]
DEBUG    │  ┕━ build merkle tree [ 21.7ms | 0.48% ] dimensions: "[8x32768]"
INFO     ┕━ prove batch opening [ 1.01s | 0.06% / 22.38% ]
INFO        ┝━ compute opened values with Lagrange interpolation [ 347ms | 7.66% ]
INFO        ┕━ FRI prover [ 663ms | 0.05% / 14.66% ]
INFO           ┝━ commit phase [ 661ms | 0.06% / 14.60% ]
DEBUG          │  ┝━ fold in matrices [ 634ms | 14.01% ] height: 32768
DEBUG          │  ┝━ build merkle tree [ 10.9ms | 0.24% ] dimensions: "[8x16384]"
DEBUG          │  ┝━ fold_even_odd [ 1.15ms | 0.03% ]
DEBUG          │  ┝━ build merkle tree [ 5.46ms | 0.12% ] dimensions: "[8x8192]"
DEBUG          │  ┝━ fold_even_odd [ 649µs | 0.01% ]
DEBUG          │  ┝━ build merkle tree [ 2.75ms | 0.06% ] dimensions: "[8x4096]"
DEBUG          │  ┝━ fold_even_odd [ 288µs | 0.01% ]
DEBUG          │  ┝━ build merkle tree [ 1.41ms | 0.03% ] dimensions: "[8x2048]"
DEBUG          │  ┝━ fold_even_odd [ 151µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 717µs | 0.02% ] dimensions: "[8x1024]"
DEBUG          │  ┝━ fold_even_odd [ 76.8µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 386µs | 0.01% ] dimensions: "[8x512]"
DEBUG          │  ┝━ fold_even_odd [ 38.5µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 205µs | 0.00% ] dimensions: "[8x256]"
DEBUG          │  ┝━ fold_even_odd [ 20.2µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 111µs | 0.00% ] dimensions: "[8x128]"
DEBUG          │  ┝━ fold_even_odd [ 10.5µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 71.2µs | 0.00% ] dimensions: "[8x64]"
DEBUG          │  ┝━ fold_even_odd [ 6.29µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 41.4µs | 0.00% ] dimensions: "[8x32]"
DEBUG          │  ┝━ fold_even_odd [ 4.21µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 22.5µs | 0.00% ] dimensions: "[8x16]"
DEBUG          │  ┝━ fold_even_odd [ 2.42µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 13.5µs | 0.00% ] dimensions: "[8x8]"
DEBUG          │  ┝━ fold_even_odd [ 1.58µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 33.9µs | 0.00% ] dimensions: "[8x4]"
DEBUG          │  ┝━ fold_even_odd [ 1.25µs | 0.00% ]
DEBUG          │  ┝━ build merkle tree [ 5.33µs | 0.00% ] dimensions: "[8x2]"
DEBUG          │  ┕━ fold_even_odd [ 1.04µs | 0.00% ]
INFO           ┕━ query phase [ 541µs | 0.01% ]
```

The LDEs are actually a tiny bit slower, surprisingly. And the `fold_in_matrices` is twice as fast?? I have no idea why, it should be the same. Please code review but I am going to investigate why it is faster.