@rjeli 2023-12-12
Oh, weird, the difference goes away with `lto = "fat"`, in which case they both take 4.8s. maybe it was just shifting symbols around and affecting icache slightly differently